src/services/monitoring/ChangeDetectionService.ts
import { ILogger } from '../../types/logger';
import { IChangeDetectionService, ChangeDetectionResult } from './types';

export class ChangeDetectionService implements IChangeDetectionService {
  private lastCheckedUrl: string | null = null;
  private logger: ILogger;

  constructor(logger: ILogger) {
    this.logger = logger;
  }

  async checkForChanges(): Promise<ChangeDetectionResult> {
    try {
      this.logger.info('Checking for content changes');
      
      // TODO: Implement Python script integration
      // Example:
      // const pythonResult = await this.executePythonScript();
      // if (pythonResult.url !== this.lastCheckedUrl) {
      //   this.lastCheckedUrl = pythonResult.url;
      //   return {
      //     hasChanged: true,
      //     currentUrl: pythonResult.url
      //   };
      // }

      return {
        hasChanged: false,
        currentUrl: this.lastCheckedUrl || undefined
      };
    } catch (error) {
      this.logger.error('Change detection failed:', error instanceof Error ? error.message : 'Unknown error');
      return { hasChanged: false };
    }
  }

  // TODO: Add method to execute Python script
  // private async executePythonScript(): Promise<{ url: string }> {
  //   // Implementation here
  // }
}